# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shourov Hub Video Downloader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SearchRequest {
  /** Search query */
  query: string;
  /** Max number of results */
  limit?: number;
}

export interface SearchResult {
  id: string;
  title: string;
  thumbnail?: string;
  duration?: number;
  platform: string;
  url: string;
  uploader?: string;
  viewCount?: number;
}

export interface SearchResults {
  results: SearchResult[];
}

export interface VideoInfoRequest {
  /** The video URL to fetch info for */
  url: string;
}

export type VideoFormatType =
  (typeof VideoFormatType)[keyof typeof VideoFormatType];

export const VideoFormatType = {
  video: "video",
  audio: "audio",
} as const;

export interface VideoFormat {
  id: string;
  label: string;
  quality: string;
  ext: string;
  filesize?: number;
  type: VideoFormatType;
}

export interface VideoInfo {
  title: string;
  thumbnail?: string;
  duration?: number;
  platform: string;
  formats: VideoFormat[];
  originalUrl: string;
}

export type DownloadRequestType =
  (typeof DownloadRequestType)[keyof typeof DownloadRequestType];

export const DownloadRequestType = {
  video: "video",
  audio: "audio",
} as const;

export interface DownloadRequest {
  url: string;
  formatId: string;
  type: DownloadRequestType;
}

export interface DownloadResult {
  downloadUrl: string;
  filename: string;
  filesize?: number;
  message?: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}