# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Shourov Hub Video Downloader API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: download
    description: Video download operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /download/search:
    post:
      operationId: searchVideos
      tags: [download]
      summary: Search videos
      description: Search for videos by keyword using YouTube search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /download/info:
    post:
      operationId: getVideoInfo
      tags: [download]
      summary: Get video info
      description: Fetch video information from a URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInfoRequest"
      responses:
        "200":
          description: Video info retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoInfo"
        "400":
          description: Invalid URL or unsupported platform
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /download/start:
    post:
      operationId: startDownload
      tags: [download]
      summary: Start download
      description: Start downloading a video
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DownloadRequest"
      responses:
        "200":
          description: Download link returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SearchRequest:
      type: object
      properties:
        query:
          type: string
          description: Search query
        limit:
          type: number
          description: Max number of results
      required:
        - query
    SearchResult:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        thumbnail:
          type: string
        duration:
          type: number
        platform:
          type: string
        url:
          type: string
        uploader:
          type: string
        viewCount:
          type: number
      required:
        - id
        - title
        - url
        - platform
    SearchResults:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
      required:
        - results
    VideoInfoRequest:
      type: object
      properties:
        url:
          type: string
          description: The video URL to fetch info for
      required:
        - url
    VideoFormat:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        quality:
          type: string
        ext:
          type: string
        filesize:
          type: number
        type:
          type: string
          enum: [video, audio]
      required:
        - id
        - label
        - quality
        - ext
        - type
    VideoInfo:
      type: object
      properties:
        title:
          type: string
        thumbnail:
          type: string
        duration:
          type: number
        platform:
          type: string
        formats:
          type: array
          items:
            $ref: "#/components/schemas/VideoFormat"
        originalUrl:
          type: string
      required:
        - title
        - platform
        - formats
        - originalUrl
    DownloadRequest:
      type: object
      properties:
        url:
          type: string
        formatId:
          type: string
        type:
          type: string
          enum: [video, audio]
      required:
        - url
        - formatId
        - type
    DownloadResult:
      type: object
      properties:
        downloadUrl:
          type: string
        filename:
          type: string
        filesize:
          type: number
        message:
          type: string
      required:
        - downloadUrl
        - filename
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message