# lib/api-zod/src/generated/types/searchRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shourov Hub Video Downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchRequest {
  /** Search query */
  query: string;
  /** Max number of results */
  limit?: number;
}