# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shourov Hub Video Downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchResult {
  id: string;
  title: string;
  thumbnail?: string;
  duration?: number;
  platform: string;
  url: string;
  uploader?: string;
  viewCount?: number;
}