# lib/api-zod/src/generated/types/searchResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shourov Hub Video Downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { SearchResult } from "./searchResult";

export interface SearchResults {
  results: SearchResult[];
}